# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handlers.LoginReq:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  handlers.LoginRes:
    properties:
      token:
        type: string
    type: object
  handlers.RegisterReq:
    properties:
      code:
        type: string
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
      specialization:
        type: string
      username:
        type: string
    type: object
  handlers.RegisterRes:
    properties:
      id:
        type: integer
    type: object
  handlers.RestoreReq:
    properties:
      code:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
...
  handlers.UpdateReq:
    properties:
      bio:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      specialization:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: Appointment booking system
  title: Strawberry API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return a JWT token
      parameters:
      - description: login credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Login user
      tags:
      - auth
...
  /register:
    post:
      consumes:
      - application/json
      description: Create a new user or master account
      parameters:
      - description: registration data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.RegisterRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Register a new user or master
      tags:
      - auth
  /restore:
    post:
      consumes:
      - application/json
      description: Verifies a code and sets a new password for the user
      parameters:
      - description: Restore request payload
        in: body
        name: restore
        required: true
        schema:
          $ref: '#/definitions/handlers.RestoreReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Invalid data or validation error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Invalid code
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Restore password
      tags:
      - auth
...
  /users:
    put:
      consumes:
      - application/json
      description: Update User's data such as username , bio,  full_name, specialization
        and email
      parameters:
      - description: update data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateReq'
      responses:
        "200":
          description: Апдейт
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update User
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"